+++ CHANGELOG.md
@@ -13,17 +13,24 @@ CHANGELOG
 * The following payment processors were added to the `Payment.Processor` enum:
   * `AFFIRM`
   * `AFTERPAY`
+  * `CETELEM`
   * `DATACASH`
   * `DOTPAY`
+  * `ECOMMPAY`
+  * `G2A_PAY`
   * `GOCARDLESS`
   * `INTERAC`
   * `KLARNA`
+  * `MERCANET`
   * `PAYEEZY`
   * `PAYLIKE`
   * `PAYMENT_EXPRESS`
   * `PAYSAFECARD`
   * `SMARTDEBIT`
   * `SYNAPSEFI`
+  * `VEREPAY`
+* Deprecated `VERAPAY` in the `Payment.Processor` enum. This was a misspelling
+  of `VEREPAY`.
 * Deprecated `getEmailTenure()` and `getIpTenure()` methods of
   `com.maxmind.minfraud.response.Subscores`.
 * Deprecated the `isHighRisk()` method of `com.maxmind.minfraud.response.GeoIP2Country`.